Rename image_avif folder to libavifinfo
Fix Windows build.
Revert config.m4 formatting to the previous one.
