+++ ext/standard/config.m4
@@ -451,17 +451,16 @@ dnl Setup extension sources
 dnl
 PHP_NEW_EXTENSION(standard, array.c base64.c basic_functions.c browscap.c crc32.c crypt.c \
                             datetime.c dir.c dl.c dns.c exec.c file.c filestat.c \
-                            flock_compat.c formatted_print.c fsock.c head.c html.c \
-                            image.c image_avif/avifinfo.c info.c iptc.c lcg.c link.c \
-                            mail.c math.c md5.c metaphone.c microtime.c pack.c pageinfo.c \
-                            quot_print.c rand.c mt_rand.c soundex.c string.c scanf.c \
-                            syslog.c type.c uniqid.c url.c var.c versioning.c assert.c \
-                            strnatcmp.c levenshtein.c incomplete_class.c url_scanner_ex.c \
-                            ftp_fopen_wrapper.c http_fopen_wrapper.c php_fopen_wrapper.c \
-                            credits.c css.c var_unserializer.c ftok.c sha1.c \
-                            user_filters.c uuencode.c filters.c proc_open.c \
-                            streamsfuncs.c http.c password.c random.c net.c hrtime.c \
-                            crc32_x86.c,,,
+                            flock_compat.c formatted_print.c fsock.c head.c html.c image.c \
+                            info.c iptc.c lcg.c link.c mail.c math.c md5.c metaphone.c \
+                            microtime.c pack.c pageinfo.c quot_print.c rand.c mt_rand.c \
+                            soundex.c string.c scanf.c syslog.c type.c uniqid.c url.c \
+                            var.c versioning.c assert.c strnatcmp.c levenshtein.c \
+                            incomplete_class.c url_scanner_ex.c ftp_fopen_wrapper.c \
+                            http_fopen_wrapper.c php_fopen_wrapper.c credits.c css.c \
+                            var_unserializer.c ftok.c sha1.c user_filters.c uuencode.c \
+                            filters.c proc_open.c streamsfuncs.c http.c password.c \
+                            random.c net.c hrtime.c crc32_x86.c libavifinfo/avifinfo.c,,,
 			    -DZEND_ENABLE_STATIC_TSRMLS_CACHE=1)
 
 PHP_ADD_MAKEFILE_FRAGMENT
