+++ ext/standard/libavifinfo/avifinfo.c
