+++ ext/standard/libavifinfo/avifinfo.h
